% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableDownload.R
\name{table_download}
\alias{table_download}
\title{table_download}
\usage{
table_download(table, geography, source = "acs", year = 2018,
  state = NULL, county = NULL, geometry = FALSE)
}
\arguments{
\item{table}{a character string either of the table name or concept (consult AllVariables)}

\item{geography}{The geography of your data, e.g. "tract", "county", "state"}

\item{source}{either "acs" or "sf1"}

\item{year}{one of 1990, 2000, or 2010 for decennial data, or 2010-2018 for ACS data}

\item{state}{an optional state specification, names, postal abbreviations, and FIPS accepted}

\item{county}{an optional county specification, names and FIPS accepted}

\item{geometry}{logical, if TRUE returns sf object}
}
\value{
a dataframe
}
\description{
table_download
}
\examples{
table_download("HISPANIC OR LATINO ORIGIN BY RACE", geography = "tract", source = "acs", year = 2018, state = "WI", county = "079")
}
